+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.0 (IN PROGRESS)
 * [STHUB-1](https://folio-org.atlassian.net/browse/STHUB-1) Added config and logic for session validation. If session is invalid, redirect to keycloak login page.
-* [STHUB-18](https://folio-org.atlassian.net/browse/STHUB-18) Store stripes-config and branding data in localforage to override stripes.config.js values (or absence of stripes.config.js).
+* [STHUB-18](https://folio-org.atlassian.net/browse/STHUB-18) Store stripes-config and branding data in localStorage to override stripes.config.js values (or absence of stripes.config.js).
 
 ## 1.0.0